webkit/data/layout_test_results/v8/LayoutTests/http/tests/security/xss-DENIED-synchronous-form-expected.txt

CONSOLE MESSAGE: line 11: Uncaught TypeError: Object [object global] has no method 'postMessage'

Waiting for iframe...
FAIL: Cross-domain iframe was able to access this frame
